+++ .github/workflows/build-web.yml
@@ -14,7 +14,7 @@ jobs:
           fetch-depth: 0
 
       - name: Remove old "dist" folder 
-        working-directory: ./web/frontend
+        working-directory: ./frontend
         run: |
           if (Test-Path -Path ./dist) { rm -r dist }
           
@@ -25,14 +25,14 @@ jobs:
           deno-version: v1.x
 
       - name: Build Frontend 📄
-        working-directory: ./web/frontend
+        working-directory: ./frontend
         run: |
           deno run --allow-run --allow-read https://deno.land/x/snel/install.ts
           trex run build
           if (Test-Path -Path ./trex_storage) { rm -r trex_storage }
       
       - name: Copy Images 🖼️
-        working-directory: ./web/frontend
+        working-directory: ./frontend
         run: |
           copy -r .\public\images\ .\dist\
