+++ backend/controller/matchController.ts
@@ -1,17 +1,14 @@
-import { Participant, ParticipantSchema } from "../db/schemas/participant.ts";
-import { Bson, Context } from "../../../deps/web/deps.ts";
+import { Matches, MatchesSchema } from "../db/schemas/matches.ts";
+import { Bson, Context } from "../../deps/web/deps.ts";
 
-// creates a new participant with data in request
-const createParticipant = async (ctx: Context) => {
+const createMatches = async (ctx: Context) => {
   try {
     const body = await ctx.request.body();
-    const { name, gender, age, img, season } = await body.value;
-    const data = await Participant.insertOne({
-      name: name,
-      age: age,
-      img: img,
+    const { matches, season, episode } = await body.value;
+    const data = await Matches.insertOne({
+      matches: matches,
       season: season,
-      gender,
+      episode: episode,
     });
     ctx.response.body = { status: true, data: data };
     ctx.response.status = 201;
@@ -25,11 +22,9 @@ const createParticipant = async (ctx: Context) => {
   }
 };
 
-// returns all participants from db
-const getParticipants = async (ctx: Context) => {
+const getMatches = async (ctx: Context) => {
   try {
-    const allParticipant: ParticipantSchema[] = await Participant.find({})
-      .toArray();
+    const allParticipant: MatchesSchema[] = await Matches.find({}).toArray();
     ctx.response.body = { status: true, data: allParticipant };
     ctx.response.status = 200;
   } catch (error) {
@@ -42,14 +37,17 @@ const getParticipants = async (ctx: Context) => {
   }
 };
 
-const getParticipant = async (
+// any used in here
+const getMatch = async (
   // deno-lint-ignore no-explicit-any
   { params, response }: { params: { id: string }; response: any }, // skipcq: JS-0323
 ) => {
   try {
     const id = params.id;
     const betterId = new Bson.ObjectId(id);
-    const participant = await Participant.findOne({ _id: betterId });
+    const participant: MatchesSchema | undefined = await Matches.findOne({
+      _id: betterId,
+    });
     response.body = { status: true, data: participant };
     response.status = 200;
   } catch (error) {
@@ -62,4 +60,4 @@ const getParticipant = async (
   }
 };
 
-export { createParticipant, getParticipant, getParticipants };
+export { createMatches, getMatch, getMatches };
