+++ backend/controller/participantController.ts
@@ -1,14 +1,17 @@
-import { Matches, MatchesSchema } from "../db/schemas/matches.ts";
-import { Bson, Context } from "../../../deps/web/deps.ts";
+import { Participant, ParticipantSchema } from "../db/schemas/participant.ts";
+import { Bson, Context } from "../../deps/web/deps.ts";
 
-const createMatches = async (ctx: Context) => {
+// creates a new participant with data in request
+const createParticipant = async (ctx: Context) => {
   try {
     const body = await ctx.request.body();
-    const { matches, season, episode } = await body.value;
-    const data = await Matches.insertOne({
-      matches: matches,
+    const { name, gender, age, img, season } = await body.value;
+    const data = await Participant.insertOne({
+      name: name,
+      age: age,
+      img: img,
       season: season,
-      episode: episode,
+      gender,
     });
     ctx.response.body = { status: true, data: data };
     ctx.response.status = 201;
@@ -22,9 +25,11 @@ const createMatches = async (ctx: Context) => {
   }
 };
 
-const getMatches = async (ctx: Context) => {
+// returns all participants from db
+const getParticipants = async (ctx: Context) => {
   try {
-    const allParticipant: MatchesSchema[] = await Matches.find({}).toArray();
+    const allParticipant: ParticipantSchema[] = await Participant.find({})
+      .toArray();
     ctx.response.body = { status: true, data: allParticipant };
     ctx.response.status = 200;
   } catch (error) {
@@ -37,17 +42,14 @@ const getMatches = async (ctx: Context) => {
   }
 };
 
-// any used in here
-const getMatch = async (
+const getParticipant = async (
   // deno-lint-ignore no-explicit-any
   { params, response }: { params: { id: string }; response: any }, // skipcq: JS-0323
 ) => {
   try {
     const id = params.id;
     const betterId = new Bson.ObjectId(id);
-    const participant: MatchesSchema | undefined = await Matches.findOne({
-      _id: betterId,
-    });
+    const participant = await Participant.findOne({ _id: betterId });
     response.body = { status: true, data: participant };
     response.status = 200;
   } catch (error) {
@@ -60,4 +62,4 @@ const getMatch = async (
   }
 };
 
-export { createMatches, getMatch, getMatches };
+export { createParticipant, getParticipant, getParticipants };
