+++ deps/web/Procfile
@@ -1,1 +1,1 @@
-web: deno run --allow-net --allow-read --allow-env web/backend/server.ts --port=${PORT}
+web: deno run --allow-net --allow-read --allow-env backend/server.ts --port=${PORT}
