+++ discord/main.ts
@@ -10,19 +10,19 @@ import {
   slash,
 } from "../deps/discord/deps.ts";
 import { commands } from "./commands/commands.ts";
-import { db } from "../web/backend/db/mongo.ts";
+import { db } from "../backend/db/mongo.ts";
 import { placeBet } from "./commands/placeBet.ts";
 import { showBet } from "./commands/showBet.ts";
 import { showMatches } from "./commands/showMatches.ts";
 import { addMatchNight } from "./commands/addMatchNight.ts";
 import { addMatchBox } from "./commands/addMatchBox.ts";
-import { Admin, AdminSchema } from "../web/backend/db/schemas/admin.ts";
+import { Admin, AdminSchema } from "../backend/db/schemas/admin.ts";
 import { getAdminIds } from "./commands/util.ts";
 import { hotOrNot } from "./commands/hotOrNot.ts";
 import {
   Participant,
   ParticipantSchema,
-} from "../web/backend/db/schemas/participant.ts";
+} from "../backend/db/schemas/participant.ts";
 
 const env = config();
 const token = Deno.env.get("BOT_TOKEN") || env.BOT_TOKEN;
