+++ frontend/public/favicons/browserconfig.xml
