+++ frontend/public/favicons/favicon.ico
