+++ frontend/public/favicons/site.webmanifest
