+++ frontend/src/components/Home.svelte
