--- web/frontend/.gitignore
@@ -1,2 +1,0 @@
-/public/dist/
-.DS_Store
